Added configuration input file support: Added the capability to provide run configuration information through the library config4cpp. Further information about this package can be found at http://config4star.org/. Many other corrections were made to the code; the PlasmaGrid.cpp 'p' option for Magnum-PSI was corrected to read Potential rather than Pressure. The Models were updated to expect single precision (float) as accuracy vectors. The make file has been updated to build and link the library. Finally, the breakup process has been disabled and the DTOKSWell charging model has been simplified for testing processes.

+++ Models/ChargingModel.h
@@ -23,9 +23,9 @@ class ChargingModel : public Model{
 	public:
 		// Constructors
 		ChargingModel();
-		ChargingModel(std::string filename, double accuracy, std::array<bool,NumModels> models, 
+		ChargingModel(std::string filename, float accuracy, std::array<bool,NumModels> models, 
 				Matter *& sample, PlasmaData *& pdata);
-		ChargingModel(std::string filename, double accuracy, std::array<bool,NumModels> models, 
+		ChargingModel(std::string filename, float accuracy, std::array<bool,NumModels> models, 
 				Matter *& sample, PlasmaGrid & pgrid);
 
 		void CreateFile(std::string filename);
